Update mbox.rb - check if sqlite3 db exists
Fixed a couple of paths, moved the initialization of the sqlite3 db to a separate function, which first checks if the db already exists. Also changed SQL for creating indices to check if they exist. Added check for whether a file entry in the "messages" subdir list is a directory file, and skip it if it is.